requirements.txt
requests
django==2.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt
pytest==6.2.4
pytest-django==4.4.0
python-dotenv==0.13.0
pytest-pythonpath==0.7.3
django-filter